Replace the usage of org.json with new simple json parser in tests

+++ tests/org.eclipse.rap.rwt.test/src/org/eclipse/rap/rwt/widgets/BrowserUtil_Test.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2012 EclipseSource and others.
+ * Copyright (c) 2012, 2013 EclipseSource and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -124,7 +124,7 @@ public class BrowserUtil_Test {
 
     Fixture.readDataAndProcessAction( browser );
 
-    assertEquals( Integer.valueOf( 5 ), log[ 0 ] );
+    assertEquals( Double.valueOf( 5 ), log[ 0 ] );
   }
 
   @Test
